+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.0](https://github.com/cesan3/salt-ext-elasticsearch/compare/1.0.0...1.1.0) - 2023-11-17
+- Fix returner for ES8 by @cesan3 in https://github.com/cesan3/salt-ext-elasticsearch/pull/18
+
+## [1.0.0](https://github.com/cesan3/salt-ext-elasticsearch/compare/0.0.8...1.0.0) - 2023-11-08
+- Merge upstream by @cesan3 in https://github.com/cesan3/salt-ext-elasticsearch/pull/16
+
+## [0.0.8](https://github.com/cesan3/salt-ext-elasticsearch/compare/0.0.7...0.0.8) - 2023-11-06
+- fix: add elasticsearch returner by @cesan3 in https://github.com/cesan3/salt-ext-elasticsearch/pull/14
+
 ## [0.0.7](https://github.com/cesan3/salt-ext-elasticsearch/compare/0.0.6...0.0.7) - 2023-11-06
 - Revert "fix: obsoletes saltext.elasticsearch8" by @cesan3 in https://github.com/cesan3/salt-ext-elasticsearch/pull/12